FORTRAN_lib: bess*1.f90: fixed some sp/dp issues
git-svn-id: https://svn.ufz.de/svn/chs-svn/FORTRAN_chs_lib@2069 20e3f476-f421-4892-ba25-73c5d2d2c40b

+++ nr_chs/bessk1.f90
@@ -5,13 +5,13 @@ FUNCTION bessk1_s(x)
   IMPLICIT NONE
   REAL(SP), INTENT(IN) :: x
   REAL(SP) :: bessk1_s
-  REAL(DP) :: y
-  REAL(DP), DIMENSION(7) :: p = (/1.0_dp,0.15443144_dp,&
-       -0.67278579_dp,-0.18156897_dp,-0.1919402e-1_dp,&
-       -0.110404e-2_dp,-0.4686e-4_dp/)
-  REAL(DP), DIMENSION(7) :: q = (/1.25331414_dp,0.23498619_dp,&
-       -0.3655620e-1_dp,0.1504268e-1_dp,-0.780353e-2_dp,&
-       0.325614e-2_dp,-0.68245e-3_dp/)
+  REAL(sp) :: y
+  REAL(sp), DIMENSION(7) :: p = (/1.0_sp,0.15443144_sp,&
+       -0.67278579_sp,-0.18156897_sp,-0.1919402e-1_sp,&
+       -0.110404e-2_sp,-0.4686e-4_sp/)
+  REAL(sp), DIMENSION(7) :: q = (/1.25331414_sp,0.23498619_sp,&
+       -0.3655620e-1_sp,0.1504268e-1_sp,-0.780353e-2_sp,&
+       0.325614e-2_sp,-0.68245e-3_sp/)
   call assert(x > 0.0, 'bessk1_s arg')
   if (x <= 2.0) then
      y=x*x/4.0_sp
@@ -30,14 +30,14 @@ FUNCTION bessk1_v(x)
   IMPLICIT NONE
   REAL(SP), DIMENSION(:), INTENT(IN) :: x
   REAL(SP), DIMENSION(size(x)) :: bessk1_v
-  REAL(DP), DIMENSION(size(x)) :: y
+  REAL(sp), DIMENSION(size(x)) :: y
   LOGICAL(LGT), DIMENSION(size(x)) :: mask
-  REAL(DP), DIMENSION(7) :: p = (/1.0_dp,0.15443144_dp,&
-       -0.67278579_dp,-0.18156897_dp,-0.1919402e-1_dp,&
-       -0.110404e-2_dp,-0.4686e-4_dp/)
-  REAL(DP), DIMENSION(7) :: q = (/1.25331414_dp,0.23498619_dp,&
-       -0.3655620e-1_dp,0.1504268e-1_dp,-0.780353e-2_dp,&
-       0.325614e-2_dp,-0.68245e-3_dp/)
+  REAL(sp), DIMENSION(7) :: p = (/1.0_sp,0.15443144_sp,&
+       -0.67278579_sp,-0.18156897_sp,-0.1919402e-1_sp,&
+       -0.110404e-2_sp,-0.4686e-4_sp/)
+  REAL(sp), DIMENSION(7) :: q = (/1.25331414_sp,0.23498619_sp,&
+       -0.3655620e-1_sp,0.1504268e-1_sp,-0.780353e-2_sp,&
+       0.325614e-2_sp,-0.68245e-3_sp/)
   call assert(all(x > 0.0), 'bessk1_v arg')
   mask = (x <= 2.0)
   where (mask)
